fix datatype declarations

+++ 64bit_integer/include/utils.h
@@ -7,6 +7,6 @@
 
 #endif
 
-uint16_t *build_active_bits_table();
+uint8_t *build_active_bits_table();
 uint64_t bincoeff(uint16_t n, uint16_t k);
 uint64_t **build_subindex_tables();
